refactor: Move Header/Import templates to TemplateRenderer as private
